correct some bugs

+++ 12/9.c
@@ -3,16 +3,16 @@
 
 int main()
 {
-    float r;
-    float c;
-    float s;
-    scanf("%g", &r);
+    double r;
+    double c;
+    double s;
+    scanf("%lf", &r);
 
     c = 2 * PI * r;
     s = PI * r * r;
 
-    printf("%.3f\n", c);
-    printf("%.3f\n", s);
+    printf("%.3lf\n", c);
+    printf("%.3lf\n", s);
 
     return 0;
 }
